+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Updated hotspot share messages to include stat counts and shard names
 - Updated Slugfish Timer to be server tick based
 - Made Frog Blessing display save between sessions
+- Made the Rare Drops/Sc displays hidden while trophy or treasure fishing
 
 ### Fixes
 - Fixed sea creature glow only displaying one entity and invisible entities sometimes